Incorporated preliminary implementation review comments and broke up tests into three files (see AIDA-719).
