+++ src/test/java/com/ncc/aif/ExamplesAndValidationTest.java
@@ -24,9 +24,6 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Set;
 
 import static com.ncc.aif.AIFUtils.*;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -47,12 +44,8 @@ public class ExamplesAndValidationTest {
     static void declutterLogging() {
         // prevent too much logging from obscuring the Turtle examples which will be printed
         ((Logger) org.slf4j.LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME)).setLevel(Level.INFO);
-        seedlingValidator =
-                ValidateAIF.createForDomainOntologySource(SEEDLING_ONTOLOGY);
-        utils = new TestUtils(NAMESPACE, seedlingValidator);
+        utils = new TestUtils(NAMESPACE, ValidateAIF.createForDomainOntologySource(SEEDLING_ONTOLOGY), false);
     }
-
-    private static ValidateAIF seedlingValidator;
 
     private Model model;
     private Resource system;
@@ -98,10 +91,6 @@ public class ExamplesAndValidationTest {
 
     @Nested
     class ValidExamples {
-        ValidExamples() {
-            utils.setValidator(seedlingValidator);
-        }
-
         private final String putinDocumentEntityUri = getUri("E781167.00398");
         private final String putinResidesDocumentRelationUri = getUri("R779959.00000");
         private final String putinElectedDocumentEventUri = getUri("V779961.00010");
@@ -1107,9 +1096,6 @@ public class ExamplesAndValidationTest {
      */
     @Nested
     class InvalidExamples {
-        InvalidExamples() {
-            utils.setValidator(seedlingValidator);
-        }
 
         @Test
         void entityMissingType() {
@@ -1181,1055 +1167,6 @@ public class ExamplesAndValidationTest {
             final Resource entity = model.createResource("http://www.test.edu/entity/1");
             entity.addProperty(AidaAnnotationOntology.SYSTEM_PROPERTY, system);
             testInvalid("Invalid: missing rdf type");
-        }
-    }
-
-    /**
-     * Set of tests to show that NIST restrictions pass and fail appropriately
-     */
-    @Nested
-    class NISTExamples {
-        Resource entity;
-        Resource event;
-        Resource relation;
-        Resource entityCluster;
-        Resource eventCluster;
-        Resource relationCluster;
-        Resource typeAssertionJustification; // For use when tests create their own entities, events, and relations
-
-        NISTExamples() {
-            utils.setValidator(ValidateAIF.create(ImmutableSet.of(SEEDLING_ONTOLOGY),
-                    ValidateAIF.Restriction.NIST));
-        }
-
-        @BeforeEach
-        void setup() {
-            typeAssertionJustification = makeTextJustification(model, "NYT_ENG_20181231",
-                    42, 143, system, 0.973);
-            addSourceDocumentToJustification(typeAssertionJustification, "20181231");
-
-            entity = makeEntity(model, getEntityUri(), system);
-            markJustification(addType(entity, SeedlingOntology.Person), typeAssertionJustification);
-            event = makeEvent(model, getUri("event1"), system);
-            markJustification(addType(event, SeedlingOntology.Conflict_Attack), typeAssertionJustification);
-            relation = makeRelation(model, getUri("relation1"), system);
-            markJustification(addType(relation, SeedlingOntology.GeneralAffiliation_APORA), typeAssertionJustification);
-
-            entityCluster = makeClusterWithPrototype(model, getClusterUri(), entity, system);
-            eventCluster = makeClusterWithPrototype(model, getClusterUri(), event, system);
-            relationCluster = makeClusterWithPrototype(model, getClusterUri(), relation, system);
-        }
-
-        // Each edge justification must be represented uniformly in AIF by
-        // aida:CompoundJustification, even if only one span is provided
-        // Edges are assumed to be relation and event arguments
-        //
-        // CompoundJustification must be used only for justifications of argument assertions,
-        // and not for justifications for entities, events, or relation KEs
-        @Nested
-        class RestrictJustification {
-            @Test
-            void invalid() {
-
-                // test relation edge argument must have a compound justification
-                final Resource relationEdge = markAsArgument(model, relation,
-                        SeedlingOntology.GeneralAffiliation_APORA_Affiliate, entity, system, 1d, getAssertionUri());
-                final Resource justification = markTextJustification(model, relationEdge,
-                        "source1", 0, 4, system, 1d);
-                addSourceDocumentToJustification(justification, "source1sourceDocument");
-
-                // test event edge argument must have a compound justification
-                final Resource eventEdge = markAsArgument(model, event, SeedlingOntology.Conflict_Attack_Target,
-                        entity, system, 1.0, getAssertionUri());
-                markJustification(eventEdge, justification);
-
-                final Resource justification1 = makeTextJustification(model, "source1", 0, 4, system, 1d);
-                addSourceDocumentToJustification(justification1, "source1sourceDocument");
-                final Resource compound1 = markCompoundJustification(model,
-                        ImmutableSet.of(entity, relation, event),
-                        ImmutableSet.of(justification1),
-                        system,
-                        1d);
-
-                // test that compound justification can only be used for argument assertions
-                markJustification(entity, compound1);
-                markJustification(relation, compound1);
-                markJustification(event, compound1);
-
-                testInvalid("NIST.invalid: CompoundJustification must be used only for justifications of argument assertions");
-            }
-
-            @Test
-            void invalidCompoundJustificationWithNoJustification() {
-
-                Resource compoundJustification = model.createResource();
-                compoundJustification.addProperty(RDF.type, AidaAnnotationOntology.COMPOUND_JUSTIFICATION_CLASS);
-                markSystem(compoundJustification, system);
-
-                markConfidence(model, compoundJustification, 1.0, system);
-                testInvalid("NIST.invalid: (No justification in CompoundJustification) Exactly 1 or 2 contained" +
-                        " justifications in a CompoundJustification required for an edge");
-            }
-
-            @Test
-            void invalidCompoundJustificationWithThreeJustifications() {
-
-                // test relation argument
-                final Resource relationEdge = markAsArgument(model, relation,
-                        SeedlingOntology.GeneralAffiliation_APORA_Affiliate, entity, system, 1d);
-                final Resource justification1 = makeTextJustification(model, "source1", 0,
-                        4, system, 1d);
-                addSourceDocumentToJustification(justification1, "source1sourceDocument");
-                final Resource justification2 = makeTextJustification(model, "source2", 0,
-                        4, system, 1d);
-                addSourceDocumentToJustification(justification2, "source2sourceDocument");
-                final Resource justification3 = makeTextJustification(model, "source3", 0,
-                        4, system, 1d);
-                addSourceDocumentToJustification(justification3, "source3sourceDocument");
-
-                markCompoundJustification(model,
-                        ImmutableSet.of(relationEdge),
-                        ImmutableSet.of(justification1, justification2, justification3),
-                        system,
-                        1d);
-
-                testInvalid("NIST.invalid: (More than two justifications in CompoundJustification) " +
-                        "Exactly 1 or 2 contained justifications in a CompoundJustification required for an edge");
-            }
-
-            @Test
-            void valid() {
-
-                // test relation argument
-                final Resource relationEdge = markAsArgument(model, relation,
-                        SeedlingOntology.GeneralAffiliation_APORA_Affiliate, entity, system, 1d);
-                final Resource justification1 = makeTextJustification(model, "source1", 0, 4, system, 1d);
-                addSourceDocumentToJustification(justification1, "source1sourceDocument");
-
-                final Resource compound = markCompoundJustification(model,
-                        ImmutableSet.of(relationEdge),
-                        ImmutableSet.of(justification1),
-                        system,
-                        1d);
-
-                markJustification(relationEdge, compound);
-
-                // test event argument
-                final Resource eventEdge = markAsArgument(model, event, SeedlingOntology.Conflict_Attack_Target, entity, system, 1.0);
-
-                markJustification(eventEdge, compound);
-
-                testValid("NIST.valid: CompoundJustification must be used only for justifications of argument assertions");
-            }
-        }
-
-        // Each edge justification is limited to either one or two spans.
-        @Nested
-        class EdgeJustificationLimit {
-            @Test
-            void invalid() {
-                // test relation
-                final Resource relationEdge = markAsArgument(model, relation,
-                        SeedlingOntology.GeneralAffiliation_APORA_Affiliate, entity, system, 1d, getAssertionUri());
-                final Resource justification1 = makeTextJustification(model, "source1", 0, 4, system, 1d);
-                addSourceDocumentToJustification(justification1, "source1sourceDocument");
-                final Resource justification2 = makeTextJustification(model, "source1", 10, 14, system, 1d);
-                addSourceDocumentToJustification(justification2, "source1sourceDocument");
-                final Resource justification3 = makeTextJustification(model, "source1", 20, 24, system, 1d);
-                addSourceDocumentToJustification(justification3, "source1sourceDocument");
-                final Resource compound = markCompoundJustification(model,
-                        ImmutableSet.of(relationEdge),
-                        ImmutableSet.of(justification1, justification2, justification3),
-                        system,
-                        1d);
-
-                // test event
-                final Resource eventEdge = markAsArgument(model, event, SeedlingOntology.Conflict_Attack_Target, entity,
-                        system, 1.0, getAssertionUri());
-                markJustification(eventEdge, compound);
-
-                testInvalid("NIST.invalid: edge justification contains one or two mentions (three is too many)");
-            }
-
-
-            @Test
-            void invalidZeroSpans() {
-                // test relation
-                final Resource relationEdge = markAsArgument(model, relation,
-                        SeedlingOntology.GeneralAffiliation_APORA_Affiliate, entity, system, 1d);
-                final Resource compound = markCompoundJustification(model,
-                        ImmutableSet.of(relationEdge),
-                        ImmutableSet.of(), // no justification
-                        system,
-                        1d);
-
-                // test event
-                final Resource eventEdge = markAsArgument(model, event, SeedlingOntology.Conflict_Attack_Target, entity, system, 1.0);
-                markJustification(eventEdge, compound);
-
-                testInvalid("NIST.invalid: edge justification contains one or two mentions (zero is not enough)");
-            }
-
-            @Test
-            void valid() {
-                // test relation
-                final Resource relationEdge = markAsArgument(model, relation,
-                        SeedlingOntology.GeneralAffiliation_APORA_Affiliate, entity, system, 1d);
-                final Resource justification1 = makeTextJustification(model, "source1", 0, 4, system, 1d);
-                addSourceDocumentToJustification(justification1, "source1sourceDocument");
-                final Resource justification2 = makeTextJustification(model, "source1", 10, 14, system, 1d);
-                addSourceDocumentToJustification(justification2, "source1sourceDocument");
-                final Resource compound = markCompoundJustification(model,
-                        ImmutableSet.of(relationEdge),
-                        ImmutableSet.of(justification1, justification2),
-                        system,
-                        1d);
-
-                // test event
-                final Resource eventEdge = markAsArgument(model, event, SeedlingOntology.Conflict_Attack_Target, entity, system, 1.0);
-                markJustification(eventEdge, compound);
-
-                testValid("NIST.valid: edge justification contains two mentions (i.e., one or two are valid)");
-            }
-
-            @Test
-            void validOneSpan() {
-                // test relation
-                final Resource relationEdge = markAsArgument(model, relation,
-                        SeedlingOntology.GeneralAffiliation_APORA_Affiliate, entity, system, 1d);
-                final Resource justification1 = makeTextJustification(model, "source1", 0, 4, system, 1d);
-                addSourceDocumentToJustification(justification1, "source1sourceDocument");
-                final Resource compound = markCompoundJustification(model,
-                        ImmutableSet.of(relationEdge),
-                        ImmutableSet.of(justification1),
-                        system,
-                        1d);
-
-                // test event
-                final Resource eventEdge = markAsArgument(model, event, SeedlingOntology.Conflict_Attack_Target, entity, system, 1.0);
-                markJustification(eventEdge, compound);
-
-                testValid("NIST.valid: edge justification contains one mention (i.e., one or two are valid)");
-            }
-        }
-
-        // Video must use aida:KeyFrameVideoJustification. Remove ShotVideoJustification
-        @Nested
-        class PreventShotVideo {
-            @Test
-            void invalid() {
-                final Resource markShotVideoJustification = markShotVideoJustification(model, entity, "source1",
-                        "shotId", system, 1d);
-                addSourceDocumentToJustification(markShotVideoJustification, "source1SourceDocument");
-                testInvalid("NIST.invalid: No shot video");
-            }
-
-            @Test
-            void valid() {
-                final Resource markKeyFrameVideoJustification = markKeyFrameVideoJustification(model, entity,
-                        "source1", "keyframe",
-                        new BoundingBox(new Point(0, 0), new Point(100, 100)), system, 1d);
-                addSourceDocumentToJustification(markKeyFrameVideoJustification, "source1SourceDocument");
-                testValid("NIST.valid: No shot video");
-            }
-        }
-
-        // Members of clusters are entity objects, relation objects, and event objects (not clusters)
-        @Nested
-        class FlatClusters {
-            @Test
-            void invalid() {
-                markAsPossibleClusterMember(model, eventCluster, entityCluster, .5, system);
-                testInvalid("NIST.invalid: Flat clusters");
-            }
-
-            @Test
-            void valid() {
-                final Resource newEntity = makeEntity(model, getEntityUri(), system);
-                markJustification(addType(newEntity, SeedlingOntology.Person), typeAssertionJustification);
-                markAsPossibleClusterMember(model, newEntity, entityCluster, .75, system);
-                testValid("NIST.valid: Flat clusters");
-            }
-        }
-
-        // Entity, Relation, and Event object is required to be part of at least one cluster.
-        // This is true even if there is nothing else in the cluster
-        @Nested
-        class EverythingClustered {
-            @Test
-            void invalid() {
-                // Test entity, relation, and event. Correct other than being clustered
-                final Resource newEntity = makeEntity(model, getEntityUri(), system);
-                markJustification(addType(newEntity, SeedlingOntology.Weapon), typeAssertionJustification);
-                final Resource newRelation = makeRelation(model, getUri("relationX"), system);
-                markJustification(addType(newRelation, SeedlingOntology.GeneralAffiliation_APORA), typeAssertionJustification);
-                final Resource newEvent = makeEvent(model, getUri("eventX"), system);
-                markJustification(addType(newEvent, SeedlingOntology.Life_BeBorn), typeAssertionJustification);
-
-                testInvalid("NIST.invalid: Everything has cluster");
-            }
-
-            @Test
-            void valid() {
-                // setup() already creates an entity, event, and relation
-                // object, and adds them to a cluster.
-
-                testValid("NIST.valid: Everything has cluster");
-            }
-        }
-
-        // Each confidence value must be between 0 and 1
-        @Nested
-        class ConfidenceValueRange {
-            @Test
-            void invalid() {
-                final Resource newEntity = makeEntity(model, getEntityUri(), system);
-                markJustification(addType(newEntity, SeedlingOntology.Person), typeAssertionJustification);
-                markAsPossibleClusterMember(model, newEntity, entityCluster, 1.2, system);
-                testInvalid("NIST.invalid: confidence must be between 0 and 1");
-            }
-
-            @Test
-            void valid() {
-                final Resource newEntity = makeEntity(model, getEntityUri(), system);
-                markJustification(addType(newEntity, SeedlingOntology.Person), typeAssertionJustification);
-                markAsPossibleClusterMember(model, newEntity, entityCluster, .7, system);
-                testValid("NIST.valid: confidence must be between 0 and 1");
-            }
-        }
-
-        // Entity, Relation, and Event clusters must have IRI
-        @Nested
-        class ClusterHasIRI {
-            @Test
-            void invalid() {
-                // Test entity, relation, and event. Correct other than being clustered
-                makeClusterWithPrototype(model, null, entity, system);
-                makeClusterWithPrototype(model, null, relation, system);
-                makeClusterWithPrototype(model, null, event, system);
-
-                testInvalid("NIST.invalid: Cluster has IRI");
-            }
-
-            @Test
-            void valid() {
-                // setup() already creates an entity, relation, and event in clusters with an IRI.
-                testValid("NIST.valid: Cluster has IRI");
-            }
-        }
-
-        // Each entity/relation/event type statement must have at least one justification
-        @Nested
-        class JustifyTypeAssertions {
-            @Test
-            void invalid() {
-                // Create an entity, but do not mark its type assertion with a justification.
-                final Resource newEntity = makeEntity(model, getEntityUri(), system);
-                addType(newEntity, SeedlingOntology.Person);
-                makeClusterWithPrototype(model, getClusterUri(), newEntity, system);
-
-                // Create an event, but do not mark its type assertion with a justification.
-                final Resource newEvent = makeEvent(model, getUri("event-5"), system);
-                addType(newEvent, SeedlingOntology.Conflict_Attack);
-                makeClusterWithPrototype(model, getClusterUri(), newEvent, system);
-
-                // Create a relation, but do not mark its type assertion with a justification.
-                final Resource newRelation = makeRelation(model, getUri("relation-2"), system);
-                addType(newRelation, SeedlingOntology.GeneralAffiliation_APORA);
-                makeClusterWithPrototype(model, getClusterUri(), newRelation, system);
-
-                testInvalid("NIST.invalid: type assertions must be justified");
-            }
-
-            @Test
-            void valid() {
-                // setup() already makes type assertions on the entity, event,
-                // and relation objects, justified by a text justification.
-
-                testValid("NIST.valid: type assertions must be justified");
-            }
-        }
-
-        // Each entity/filler name string is limited to 256 UTF-8 characters
-        @Nested
-        class NameMaxLength {
-            @Test
-            void invalid() {
-                // assign alternate names to the entity that are longer that 256 characters.
-                markName(entity, "This is a test string that will be used to validate " +
-                        "that entity names and fillers are limited to 256 characters. This string should " +
-                        "fail because this string is exactly 257 characters long. This is filler text to " +
-                        "get to the two hundred and fifty-seven limit.");
-
-                testInvalid("NIST.invalid: Each entity name string is limited to 256 UTF-8 characters");
-            }
-
-            @Test
-            void valid() {
-                // assign alternate names to the entity that are equal and less than 256 characters.
-                markName(entity, "This is a test string that will be used to validate that entity " +
-                        "names and fillers are limited to 256 characters. This string should pass because " +
-                        "this string is exactly 256 characters long. Characters to get to the two hundred " +
-                        "and fifty-six character limit.");
-
-                markName(entity, "Small string size");
-
-                testValid("NIST.valid: Each entity name string is limited to 256 UTF-8 characters");
-            }
-        }
-
-        // Justifications require a source document and a source
-        @Nested
-        class JustificationSourceAndSourceDocument {
-            Resource newJustification;
-
-            @BeforeEach
-            void setup() {
-                // create justification from scratch
-                newJustification = model.createResource();
-                newJustification.addProperty(RDF.type, AidaAnnotationOntology.TEXT_JUSTIFICATION_CLASS);
-                if (system != null) {
-                    markSystem(newJustification, system);
-                }
-
-                markConfidence(model, newJustification, 0.973, system);
-                newJustification.addProperty(AidaAnnotationOntology.START_OFFSET,
-                        model.createTypedLiteral(41));
-                newJustification.addProperty(AidaAnnotationOntology.END_OFFSET_INCLUSIVE,
-                        model.createTypedLiteral(143));
-
-                final Resource newEvent = makeEvent(model, getUri("event-2"), system);
-                markJustification(addType(newEvent, SeedlingOntology.Conflict_Attack), newJustification);
-                makeClusterWithPrototype(model, getClusterUri(), newEvent, system);
-            }
-
-            @Test
-            void invalidNoSource() {
-                // include the source document but not the source
-                addSourceDocumentToJustification(newJustification, "HC00002ZO");
-                testInvalid("NIST.invalid (missing justification source): justifications require a source document and source");
-
-            }
-
-            @Test
-            void invalidNoSourceDocument() {
-                // include the source but not the source document
-                newJustification.addProperty(AidaAnnotationOntology.SOURCE, model.createTypedLiteral("XP043002ZO"));
-                testInvalid("NIST.invalid (missing justification source document): justifications require a source document and source");
-            }
-
-            @Test
-            void valid() {
-                // include the source and source document
-                newJustification.addProperty(AidaAnnotationOntology.SOURCE, model.createTypedLiteral("XP043002ZO"));
-                addSourceDocumentToJustification(newJustification, "HC00002ZO");
-                testValid("NIST.valid: justifications require a source document and a source");
-            }
-        }
-
-        @Nested
-        class InformativeJustification {
-
-
-            @Test
-            void invalidInformativeJustificationDuplicateParentDoc() {
-
-                final Resource duplicateParentDocJustification = makeTextJustification(model, "ZM39482011",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(duplicateParentDocJustification, "20181231");
-
-                markInformativeJustification(entity, typeAssertionJustification);
-                markInformativeJustification(entity, duplicateParentDocJustification);
-
-                testInvalid("NIST.invalid: (informative justifications have same parent document) Each Cluster, " +
-                        "Entity, Event, or Relation can specify up to one informative mention per document as long " +
-                        "as each informative mention points to a different sourceDocument");
-            }
-
-            @Test
-            void validEntityInformativeJustification() {
-
-                markInformativeJustification(entity, typeAssertionJustification);
-                testValid("NIST.valid: (entity) Each Cluster, Entity, Event, or Relation can specify up to one " +
-                        "informative mention per document as each informative mention points to a " +
-                        "different sourceDocument");
-            }
-
-            @Test
-            void validEventInformativeJustification() {
-                markInformativeJustification(event, typeAssertionJustification);
-                testValid("NIST.valid: (event) Each Cluster, Entity, Event, or Relation can specify up to one " +
-                        "informative mention per document as each informative mention points to a different " +
-                        "sourceDocument");
-
-            }
-
-            @Test
-            void validRelationInformativeJustification() {
-                markInformativeJustification(relation, typeAssertionJustification);
-                testValid("NIST.valid: (relation) Each Cluster, Entity, Event, or Relation can specify up to one " +
-                        "informative mention per document as each informative mention points to a different " +
-                        "sourceDocument");
-            }
-
-            @Test
-            void validClusterInformativeJustification() {
-                markInformativeJustification(entityCluster, typeAssertionJustification);
-                testValid("NIST.valid: (cluster) Each Cluster, Entity, Event, or Relation can specify up to one " +
-                        "informative mention per document as each informative mention points to a different " +
-                        "sourceDocument");
-            }
-
-            @Test
-            void validRelationWithMultipleInformativeJustifications() {
-
-                final Resource secondJustification = makeTextJustification(model, "EJ39281",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(secondJustification, "3822029");
-
-                final Resource thirdJustification = makeTextJustification(model, "CL33838",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(thirdJustification, "3948290");
-
-                //add three informative justifications to same relation KE
-                markInformativeJustification(relation, typeAssertionJustification);
-                markInformativeJustification(relation, secondJustification);
-                markInformativeJustification(relation, thirdJustification);
-
-                testValid("NIST.valid: (multiple informative justifications on relation) Each Cluster," +
-                        "Entity, Event, or Relation can specify up to one informative mention per document as long " +
-                        "as each informative mention points to a different sourceDocument");
-
-            }
-
-            @Test
-            void validEntityClusterSeparateInformativeJustificationsWithSameParentDoc() {
-
-                final Resource duplicateParentDocJustification = makeTextJustification(model, "ZM39482011",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(duplicateParentDocJustification, "20181231");
-
-                final Resource secondEntityJustification = makeTextJustification(model, "EJ39281",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(secondEntityJustification, "3822029");
-
-                final Resource secondClusterJustification = makeTextJustification(model, "CL33838",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(secondClusterJustification, "3298329");
-
-                //add informative justification in separate KE's with same parent doc
-                markInformativeJustification(entity, typeAssertionJustification);
-                markInformativeJustification(entity, secondEntityJustification);
-
-                //add more than on informative justification to the KE's
-                markInformativeJustification(entityCluster, duplicateParentDocJustification);
-                markInformativeJustification(entityCluster, secondClusterJustification);
-
-                testValid("NIST.valid: (Two KE's with informative justifications with same parent doc) Each " +
-                        "Cluster, Entity, Event, or Relation can specify up to one informative mention per document " +
-                        "as long as each informative mention points to a different sourceDocument");
-
-            }
-        }
-
-        @Nested
-        class LinkAssertion {
-            Resource linkAssertion;
-
-            void link(Resource toLink) {
-                toLink.addProperty(AidaAnnotationOntology.LINK, linkAssertion);
-            }
-
-            void target(String externalKbId) {
-                linkAssertion.addProperty(AidaAnnotationOntology.LINK_TARGET, model.createTypedLiteral(externalKbId));
-            }
-
-            @BeforeEach
-            void setup() {
-                model.removeAll();
-                system = makeSystemWithURI(model, getTestSystemUri());
-
-                linkAssertion = model.createResource();
-                linkAssertion.addProperty(RDF.type, AidaAnnotationOntology.LINK_ASSERTION_CLASS);
-                markSystem(linkAssertion, system);
-
-                typeAssertionJustification = makeTextJustification(model, "NYT_ENG_20181231",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(typeAssertionJustification, "20181231");
-
-                entity = makeEntity(model, getEntityUri(), system);
-                markJustification(addType(entity, SeedlingOntology.Person), typeAssertionJustification);
-
-                entityCluster = makeClusterWithPrototype(model, getClusterUri(), entity, system);
-            }
-
-            @Test
-            void invalidLinkToNonAssertion() {
-                linkAssertion.listProperties().toList().forEach(model::remove);
-                entity.addProperty(AidaAnnotationOntology.LINK, typeAssertionJustification);
-                testInvalid("LinkAssertion.invalid: Link to non-LinkAssertion");
-            }
-            @Test
-            void invalidNoTarget() {
-                link(entity);
-                markConfidence(model, linkAssertion, 1d, system);
-                testInvalid("LinkAssertion.invalid: No link target");
-            }
-            @Test
-            void invalidTooManyTargets() {
-                link(entity);
-                markConfidence(model, linkAssertion, 1d, system);
-                target("SomeExternalKBId-1");
-                target("SomeExternalKBId-2");
-                testInvalid("LinkAssertion.invalid: Too many link targets");
-            }
-            @Test
-            void invalidNoConfidence() {
-                link(entity);
-                target("SomeExternalKBId-1");
-                testInvalid("LinkAssertion.invalid: No confidence");
-            }
-            @Test
-            void invalidTooManyConfidences() {
-                link(entity);
-                markConfidence(model, linkAssertion, 1d, system);
-                markConfidence(model, linkAssertion, .5, system);
-                target("SomeExternalKBId-1");
-                testInvalid("LinkAssertion.invalid: Too many confidences");
-            }
-            @Test
-            void valid() {
-                link(entity);
-                markConfidence(model, linkAssertion, 1d, system);
-                target("SomeExternalKBId-1");
-                testValid("LinkAssertion.valid");
-            }
-        }
-    }
-
-    /**
-     * Set of tests to show that NIST Hypothesis restrictions pass and fail appropriately
-     */
-    @Nested
-    class NISTHypothesisExamples {
-        Resource entity;
-        Resource entityCluster;
-        Resource event;
-        Resource eventEdge;
-        Resource justification;
-
-        NISTHypothesisExamples() {
-            utils.setValidator(ValidateAIF.create(ImmutableSet.of(SEEDLING_ONTOLOGY),
-                    ValidateAIF.Restriction.NIST_HYPOTHESIS));
-        }
-
-        @BeforeEach
-        void setup() {
-            justification = makeTextJustification(model, "NYT_ENG_20181231",
-                    42, 143, system, 0.973);
-            addSourceDocumentToJustification(justification, "20181231");
-
-            entity = makeEntity(model, getEntityUri(), system);
-            entityCluster = makeClusterWithPrototype(model, getClusterUri(), entity, "handle", system);
-            markJustification(addType(entity, SeedlingOntology.Person), justification);
-
-            event = makeEvent(model, getUri("event-1"), system);
-            markImportance(makeClusterWithPrototype(model, getClusterUri(), event, "Event", system), 104.0);
-            markJustification(addType(event, SeedlingOntology.Conflict_Attack), justification);
-
-            eventEdge = markAsArgument(model, event, SeedlingOntology.Conflict_Attack_Attacker,
-                    entity, system, 1d, getAssertionUri());
-            markImportance(eventEdge, 101.0);
-        }
-
-        private Resource makeValidHypothesis(Resource... resources) {
-            Set<Resource> set = new HashSet<>();
-            Collections.addAll(set, resources);
-            final Resource hypothesis = makeHypothesis(model, getUri("hypothesis-1"), set, system);
-            markImportance(hypothesis, 100.0);
-            return hypothesis;
-        }
-
-        // Exactly 1 hypothesis should exist in model
-        @Nested
-        class SingleHypothesis {
-
-            @Test
-            void invalidTooMany() {
-                makeValidHypothesis(entity, event, eventEdge);
-                markImportance(makeHypothesis(model, getUri("hypothesis-2"),
-                        ImmutableSet.of(entity, event, eventEdge), system), 101.0);
-                testInvalid("NISTHypothesis.invalid (too many): there should be exactly 1 hypothesis");
-            }
-
-            @Test
-            void invalidTooFew() {
-                testInvalid("NISTHypothesis.invalid (too few): there should be exactly 1 hypothesis");
-            }
-
-            @Test
-            void valid() {
-                makeValidHypothesis(entity, event, eventEdge);
-                testValid("NISTHypothesis.valid: there should be exactly 1 hypothesis");
-            }
-        }
-
-        // Each entity (cluster) in the hypothesis graph must have exactly one handle
-        @Nested
-        class EntityClusterRequiredHandle {
-
-            @Test
-            // No handle property on entity cluster in hypothesis
-            void invalidNoHandle() {
-                final Resource newEntity = makeEntity(model, getEntityUri(), system);
-                makeClusterWithPrototype(model, getClusterUri(), newEntity, system);
-                markJustification(addType(newEntity, SeedlingOntology.Person), justification);
-                makeValidHypothesis(entity, newEntity, event, eventEdge);
-
-                testInvalid("NISTHypothesis.invalid (no handle exists): Each entity cluster in the hypothesis " +
-                        "graph must have exactly one handle");
-            }
-
-            @Test
-            // Two handle properties on entity cluster in hypothesis
-            void invalidMultipleHandles() {
-
-                final Resource newEntity = makeEntity(model, getEntityUri(), system);
-                final Resource cluster = makeClusterWithPrototype(model, getClusterUri(), newEntity,
-                        "handle2", system);
-                cluster.addProperty(AidaAnnotationOntology.HANDLE, "handle3");
-                markJustification(addType(newEntity, SeedlingOntology.Person), justification);
-                makeValidHypothesis(entity, newEntity, event, eventEdge);
-
-                testInvalid("NISTHypothesis.invalid (multiple handles exist): Each entity cluster in the " +
-                        "hypothesis graph must have exactly one handle");
-            }
-
-            @Test
-            // One handle on entity cluster in hypothesis
-            void valid() {
-                makeValidHypothesis(entity, event, eventEdge);
-                testValid("NISTHypothesis.valid: Each entity cluster in the hypothesis graph must have " +
-                        "exactly one handle");
-            }
-        }
-
-        // Each hypothesis graph must have exactly one hypothesis importance value
-        @Nested
-        class HypothesisImportanceValue {
-
-            @Test
-            void invalid() {
-                //invalid hypothesis, no importance value
-                makeHypothesis(model, getUri("hypothesis-1"), ImmutableSet.of(entity, event, eventEdge), system);
-                testInvalid("NISTHypothesis.invalid (hypothesis has no importance value): Each hypothesis " +
-                        "graph must have exactly one hypothesis importance value");
-            }
-            @Test
-            void valid() {
-                makeValidHypothesis(entity, event, eventEdge);
-                testValid("NISTHypothesis.valid: Each hypothesis graph must have exactly one" +
-                        " hypothesis importance value");
-            }
-        }
-
-        // Each event or relation (cluster) in the hypothesis must have exactly one importance value
-        @Nested
-        class HypothesisEventRelationClusterImportanceValue {
-
-            private final String documentEventUri = getUri("event-1");
-            private final String relationUri = getUri("relation-1");
-            Resource relation;
-            Resource eventCluster;
-            Resource relationCluster;
-            Resource relationEdge;
-
-            @BeforeEach
-            void setup() {
-                eventCluster = makeClusterWithPrototype(model, getClusterUri(), event, system);
-
-                relation = makeRelation(model, relationUri, system);
-                markJustification(markType(model, getAssertionUri(), relation,
-                        SeedlingOntology.GeneralAffiliation_APORA, system, 1.0), justification);
-
-                relationEdge = markAsArgument(model, relation, SeedlingOntology.GeneralAffiliation_APORA_Affiliate,
-                        entity, system, 1d, getAssertionUri());
-                markImportance(relationEdge, 102.0);
-
-                relationCluster = makeClusterWithPrototype(model, getClusterUri(), relation, system);
-
-                makeValidHypothesis(entity, event, eventEdge, relation, relationEdge);
-            }
-
-            @Test
-            void invalidEvent() {
-                //invalid event cluster, no importance value
-                markImportance(relationCluster, 99.0);
-                testInvalid("NISTHypothesis.invalid (event cluster has no importance value): Each event or " +
-                        "relation (cluster) in the hypothesis must have exactly one importance value");
-            }
-
-            @Test
-            void invalidRelation() {
-                //invalid relation cluster, no importance value
-                markImportance(eventCluster, 88.0);
-                testInvalid("NISTHypothesis.invalid (relation cluster has no importance value): Each event or " +
-                        "relation (cluster) in the hypothesis must have exactly one importance value");
-            }
-
-            @Test
-            void valid() {
-                markImportance(eventCluster, 88.0);
-                markImportance(relationCluster, 99.0);
-                testValid("NISTHypothesis.valid: Each event or relation (cluster) in the hypothesis must " +
-                        "have exactly one importance value");
-            }
-        }
-
-        // Each edge KE in the hypothesis graph must have exactly one edge importance value
-        @Nested
-        class HypothesisEdgeImportanceValue {
-
-            private final String relationUri = getUri("relation-1");
-            Resource relation;
-
-            @BeforeEach
-            void setup() {
-                relation = makeRelation(model, relationUri, system);
-                markJustification(markType(model, getAssertionUri(), relation,
-                        SeedlingOntology.GeneralAffiliation_APORA, system, 1.0), justification);
-
-                //markImportance(makeClusterWithPrototype(model, getClusterUri(), event, system), 88.0);
-                markImportance(makeClusterWithPrototype(model, getClusterUri(), relation, system), 88.0);
-            }
-
-            @Test
-            void invalidEventEdge() {
-
-                //invalid event argument, needs importance value
-                Resource invalidEventEdge = markAsArgument(model, event, SeedlingOntology.Personnel_Elect_Elect,
-                        entity, system, 0.785, "event-argument-1");
-
-                makeValidHypothesis(entity, event, eventEdge, relation, invalidEventEdge);
-
-                testInvalid("NISTHypothesis.invalid (event edge has no importance value): Each edge KE in the " +
-                        "hypothesis graph must have exactly one edge importance value");
-            }
-
-            @Test
-            void invalidRelationEdge() {
-
-                //invalid relation argument, needs importance value
-                Resource invalidRelationEdge = markAsArgument(model, relation, SeedlingOntology.GeneralAffiliation_APORA_Affiliation,
-                        entity, system, 0.785, "relation-argument-1");
-
-                makeValidHypothesis(entity, event, eventEdge, relation, invalidRelationEdge);
-
-                testInvalid("NISTHypothesis.invalid (relation edge has no importance value): Each edge KE in the " +
-                        "hypothesis graph must have exactly one edge importance value");
-            }
-
-            @Test
-            void validEventEdge() {
-
-                makeValidHypothesis(entity, event, eventEdge, relation);
-
-                testValid("NISTHypothesis.valid (event edge has importance value): Each edge KE in the " +
-                        "hypothesis graph must have exactly one edge importance value");
-            }
-
-            @Test
-            void validRelationEdge() {
-
-                // link entity to the relation
-                Resource relationEdge = markAsArgument(model, relation, SeedlingOntology.GeneralAffiliation_APORA_Affiliation,
-                        entity, system, 0.785, "relation-argument-1");
-                markImportance(relationEdge, 120.0);
-
-                makeValidHypothesis(entity, event, eventEdge, relation, relationEdge);
-
-                testValid("NISTHypothesis.valid (relation edge has importance value): Each edge KE in the " +
-                        "hypothesis graph must have exactly one edge importance value");
-            }
-        }
-
-        @Nested
-        class KEsInHypothesisMustBeDefined {
-
-            @Test
-            void invalid() {
-                Resource fakeEntity = model.createResource(getEntityUri());
-                makeValidHypothesis(fakeEntity, entity, event, eventEdge);
-                testInvalid("NISTHypothesis.invalid (entity is not defined): All KEs referenced by hypothesis " +
-                        "must be defined in model");
-            }
-
-            @Test
-            void valid() {
-                makeValidHypothesis(entity, event, eventEdge);
-                testValid("NISTHypothesis.valid: All KEs referenced by hypothesis must be defined in model");
-            }
-        }
-
-        @Nested
-        class KEsInModelMustBeReferencedByHypothesis{
-            Resource relation;
-            Resource relationEdge;
-
-            @BeforeEach
-            void setup() {
-                relation = makeRelation(model, getUri("relation-1"), system);
-                markImportance(makeClusterWithPrototype(model, getClusterUri(), relation, "Relation", system), 103.0);
-                markJustification(addType(relation, SeedlingOntology.GeneralAffiliation_APORA), justification);
-
-                relationEdge = markAsArgument(model, relation, SeedlingOntology.GeneralAffiliation_APORA_Affiliate,
-                        entity, system, 1d, getAssertionUri());
-                markImportance(relationEdge, 102.0);
-            }
-
-            @Test
-            void invalid() {
-                makeValidHypothesis(entity, relation, relationEdge);
-                testInvalid("NISTHypothesis.invalid (event and event edge is not referenced in hypothesis): " +
-                        "All KEs in model must be referenced by hypothesis");
-            }
-
-            @Test
-            void valid() {
-                makeValidHypothesis(entity, relation, relationEdge, event, eventEdge);
-                testValid("NISTHypothesis.valid: All KEs in model must be referenced by hypothesis");
-            }
-        }
-
-        // Each hypothesis graph must have at least one event or relation with at least one edge.
-        @Nested
-        class HypothesisRequiredOneEventOrRelationWithOneEdge {
-
-            @Test
-            void invalid() {
-                //remove everything in the model to ensure no edge KE's exist
-                ExamplesAndValidationTest.this.setup();
-                justification = makeTextJustification(model, "NYT_ENG_20181231",
-                        42, 143, system, 0.973);
-                addSourceDocumentToJustification(justification, "NYT_PARENT_ENG_20181231_03");
-
-
-                entity = makeEntity(model, getEntityUri(), system);
-                entityCluster = makeClusterWithPrototype(model, getClusterUri(), entity, "handle", system);
-                markJustification(addType(entity, SeedlingOntology.Person), justification);
-                makeValidHypothesis(entity);
-                testInvalid("NISTHypothesis.invalid (no event or relation exists): Each hypothesis graph must " +
-                        "have at least one event or relation with at least one edge.");
-            }
-
-            @Test
-            void invalidRelationAndEventEdge() {
-                final Resource relation = makeRelation(model, getUri("relation-1"), system);
-                markImportance(makeClusterWithPrototype(model, getClusterUri(), relation, "Relation", system), 103.0);
-                markJustification(addType(relation, SeedlingOntology.GeneralAffiliation_APORA), justification);
-
-                //create invalid relation edge with event argument type
-                final Resource invalidRelationEdge = markAsArgument(model, relation, SeedlingOntology.Conflict_Attack_Attacker,
-                        entity, system, 1d, getAssertionUri());
-                markImportance(invalidRelationEdge, 102.0);
-
-                makeValidHypothesis(entity, event, eventEdge, relation, invalidRelationEdge);
-                testInvalid("NISTHypothesis.invalid (event has invalid relation edge): Each hypothesis graph " +
-                        "must have at least one event or relation with at least one edge.");
-            }
-
-            // TODO This test case needs to be updated and @Test needs to be added back in  once we decide on the
-            // TODO new design of this class upon the completion of AIDA-698.
-            void validRelationAndRelationEdge() {
-                final Resource relation = makeRelation(model, getUri("relation-1"), system);
-                markImportance(makeClusterWithPrototype(model, getClusterUri(), relation, "Relation", system), 103.0);
-                markJustification(addType(relation, SeedlingOntology.GeneralAffiliation_APORA), justification);
-
-                final Resource relationEdge = markAsArgument(model, relation, SeedlingOntology.GeneralAffiliation_APORA_Affiliate,
-                        entity, system, 1d, getAssertionUri());
-                markImportance(relationEdge, 102.0);
-
-                makeValidHypothesis(entity, event, eventEdge, relation, relationEdge);
-                testValid("NISTHypothesis.valid (relation has relation edge): Each hypothesis graph must have " +
-                        "at least one event or relation with at least one edge.");
-            }
-
-            @Test
-            void validEventAndEventEdge() {
-
-                makeValidHypothesis(entity, event, eventEdge);
-                testValid("NISTHypothesis.valid (event has event edge): Each hypothesis graph must have " +
-                        "at least one event or relation with at least one edge.");
-            }
-
-            @Test
-            void validEventRelationAndEventRelationEdge() {
-
-                final Resource relation = makeRelation(model, getUri("relation-1"), system);
-                markImportance(makeClusterWithPrototype(model, getClusterUri(), relation, "Relation", system), 103.0);
-                markJustification(addType(relation, SeedlingOntology.GeneralAffiliation_APORA), justification);
-
-                final Resource relationEdge = markAsArgument(model, relation, SeedlingOntology.GeneralAffiliation_APORA_Affiliate,
-                        entity, system, 1d, getAssertionUri());
-                markImportance(relationEdge, 102.0);
-
-                makeValidHypothesis(entity, event, eventEdge, relation, relationEdge);
-                testValid("NISTHypothesis.valid (event has event edge and relation has relation edge): Each " +
-                        "hypothesis graph must have at least one event or relation with at least one edge.");
-            }
-        }
-
-        // Clusters must be homogeneous by base class (Entity, Event, or Relation)
-        @Nested
-        class HypothesisClustersMustBeHomogeneous {
-            Resource relation;
-            Resource relationEdge;
-            Resource relationCluster;
-
-            @BeforeEach
-            void setup() {
-                relation = makeRelation(model, getUri("relation-1"), system);
-
-                relationCluster = makeClusterWithPrototype(model, getClusterUri(), relation,
-                        "Relation", system);
-
-                markImportance(relationCluster, 103.0);
-                markJustification(addType(relation, SeedlingOntology.GeneralAffiliation_APORA), justification);
-
-                relationEdge = markAsArgument(model, relation, SeedlingOntology.GeneralAffiliation_APORA_Affiliate,
-                        entity, system, 1d, getAssertionUri());
-                markImportance(relationEdge, 102.0);
-            }
-
-            @Test
-            void invalid() {
-
-                // create event cluster member to add to relation cluster
-                final Resource eventMember = makeEvent(model, getUri("event-member-1"), system);
-                markJustification(addType(eventMember, SeedlingOntology.Conflict_Attack), justification);
-
-                //add invalid event cluster member to relation cluster
-                markAsPossibleClusterMember(model, eventMember, relationCluster, 1d, system);
-
-                makeValidHypothesis(entity, event, eventEdge, eventMember, relation, relationEdge);
-                testInvalid("NISTHypothesis.invalid (event exists in relation cluster): Clusters must be " +
-                        "homogeneous by base class (Entity, Event, or Relation).");
-            }
-
-            @Test
-            void valid() {
-
-                // create relation cluster member to add to relation cluster
-                final Resource relationMember = makeRelation(model, getUri("relation-member-1"), system);
-                markJustification(addType(relationMember, SeedlingOntology.GeneralAffiliation_APORA), justification);
-
-                //add valid relation cluster member to relation cluster
-                markAsPossibleClusterMember(model, relationMember, relationCluster, 1d, system);
-
-                makeValidHypothesis(entity, event, eventEdge, relation, relationEdge, relationMember);
-                testValid("NISTHypothesis.valid: Clusters must be homogeneous by base class " +
-                        "(Entity, Event, or Relation)");
-            }
         }
     }
